added option to have raspbian_image.zip downloaded preamtively
this is added in order to enable torrent download instead of direct download from raspberrypi.org

+++ fix-ssh-on-pi.bash
@@ -70,6 +70,7 @@ image_to_download="https://downloads.raspberrypi.org/raspios_full_armhf_latest"
 url_base="https://downloads.raspberrypi.org/raspios_full_armhf/images/"
 version="$( wget -q ${url_base} -O - | awk -F '"' '/raspios_full_armhf-/ {print $8}' - | sort -nr | head -1 )"
 sha_file=$( wget -q ${url_base}/${version} -O - | awk -F '"' '/armhf-full.zip.sha256/ {print $8}' - )
+
 sha_sum=$( wget -q "${url_base}/${version}/${sha_file}" -O - | awk '{print $1}' )
 sdcard_mount="/mnt/sdcard"
 
@@ -100,7 +101,11 @@ function umount_sdcard () {
 }
 
 # Download the latest image, using the  --continue "Continue getting a partially-downloaded file"
-wget --continue ${image_to_download} -O raspbian_image.zip
+if [ ! -e raspbian_image.zip ]
+then
+	echo "Didn't find raspbian_image.zip. Going to download it now."
+	wget --continue ${image_to_download} -O raspbian_image.zip
+fi
 
 echo "Checking the SHA-1 of the downloaded image matches \"${sha_sum}\""
